Finish stake call data
